+++ src/Digdir.Domain.Dialogporten.Domain/Dialogs/Entities/DialogEntity.cs
@@ -53,12 +53,12 @@ public sealed class DialogEntity :
     /// <summary>
     /// Indicates how many transmissions are not of type Submission or Correction
     /// </summary>
-    public ushort IncomingTransmissions { get; set; }
+    public int IncomingTransmissions { get; set; }
 
     /// <summary>
     /// Indicates how many transmissions are of type Submission or Correction
     /// </summary>
-    public ushort OutgoingTransmissions { get; set; }
+    public int OutgoingTransmissions { get; set; }
 
     // === Dependent relationships ===
     public DialogStatus.Values StatusId { get; set; }
